refactoring the new delete metametrics entry to a functional componenet
