+++ ui/components/app/data-deletion-error-modal/index.ts
@@ -1,0 +1,1 @@
+export { default } from './DataDeletionErrorModal';
