+++ ui/pages/settings/security-tab/delete-metametrics-data-toggle/index.ts
@@ -1,0 +1,1 @@
+export { default } from './delete-metametrics-data-toggle';
