+++ ui/pages/settings/security-tab/metametrics-toggle/metametrics-toggle.tsx
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react';
-import { useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { useI18nContext } from '../../../../hooks/useI18nContext';
 import { MetaMetricsContext } from '../../../../contexts/metametrics';
 import {
@@ -21,6 +21,7 @@ import {
   TextColor,
   TextVariant,
 } from '../../../../helpers/constants/design-system';
+import { continueRecordingMetaMetricsData } from '../../../../ducks/app/app';
 
 const MetametricsToggle = ({
   dataCollectionForMarketing,
@@ -30,6 +31,7 @@ const MetametricsToggle = ({
   setDataCollectionForMarketing: (value: boolean) => void;
 }) => {
   const t = useI18nContext();
+  const dispatch = useDispatch();
   const trackEvent = useContext(MetaMetricsContext);
   const { enableMetametrics, error: enableMetametricsError } =
     useEnableMetametrics();
@@ -72,6 +74,7 @@ const MetametricsToggle = ({
           participateInMetaMetrics,
         },
       });
+      dispatch(continueRecordingMetaMetricsData());
     }
 
     if (dataCollectionForMarketing) {
