+++ ui/pages/settings/security-tab/security-tab.container.js
@@ -25,18 +25,11 @@ import {
 import {
   getAllNetworks,
   getIsSecurityAlertsEnabled,
-  getMetaMetricsDataDeletionDate,
-  getMetaMetricsDataDeletionStatus,
   getPetnamesEnabled,
-  getShowDeleteMetaMetricsDataModal,
-  isMetaMetricsDataDeletionMarked,
-  hasRecordedMetricsSinceDeletion,
-  getShowDataDeletionErrorModal,
 } from '../../../selectors';
 import {
   continueRecordingMetaMetricsData,
   openBasicFunctionalityModal,
-  openDeleteMetaMetricsDataModal,
   unMarkingMetaMetricsDataDeletion,
 } from '../../../ducks/app/app';
 import SecurityTab from './security-tab.component';
@@ -90,12 +83,6 @@ const mapStateToProps = (state) => {
     petnamesEnabled,
     securityAlertsEnabled: getIsSecurityAlertsEnabled(state),
     useTransactionSimulations: metamask.useTransactionSimulations,
-    showDeleteMetaMetricsDataModal: getShowDeleteMetaMetricsDataModal(state),
-    showDataDeletionErrorModal: getShowDataDeletionErrorModal(state),
-    metaMetricsDataDeletionStatus: getMetaMetricsDataDeletionStatus(state),
-    metaMetricsDataDeletionDate: getMetaMetricsDataDeletionDate(state),
-    metaMetricsDataDeletionMarked: isMetaMetricsDataDeletionMarked(state),
-    hasRecordedMetricsSinceDeletion: hasRecordedMetricsSinceDeletion(state),
   };
 };
 
@@ -133,9 +120,6 @@ const mapDispatchToProps = (dispatch) => {
     },
     setUseTransactionSimulations: (value) => {
       return dispatch(setUseTransactionSimulations(value));
-    },
-    setDeleteMetaMetricsDataModalOpen: () => {
-      return dispatch(openDeleteMetaMetricsDataModal());
     },
     updateDataDeletionTaskStatus: () => {
       return dispatch(updateDataDeletionTaskStatus());
